requirements.txt
streamlit==1.47.0
langchain==0.3.26
langchain-openai==0.3.28
fal-client==0.7.0
python-dotenv==1.1.1
pillow==11.3.0
requests==2.32.4
openai==1.97.0